+++ src/vs/workbench/contrib/accessibility/browser/accessibleViewContributions.ts
